+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+# [8.19.0] - 2025-03-??
+- Added builder and custom headers for `MemberChannel` in Conversation API
+- Added support for the following event types in Conversation API:
+  - `audio:dtmf`
+  - `audio:speaking:on`
+  - `audio:speaking:off`
+  - `member:media`
+  - `rtc:answer`
+  - `rtc:answered`
+  - `rtc:ringing`
+  - `rtc:transfer`
+  - `sip:answered`
+  - `sip:machine`
+  - `sip:amd_machine`
+  - `sip:ringing`
+
 # [8.18.0] - 2025-03-05
 - Added MMS Text, File and Content outbound message types
 - Added `getFileName` and `getFileCaption` to `com.vonage.client.messages.InboundMessage`